Fixed toHexString is not a function
For some unknown reason, the data for a job when it is run no longer contains the user _id as an ObjectId, even though when the job was added to the queue, it was an ObjectId. Instead it emerges as a string. To fix, just convert back to ObjectId before we do anything else with the job.

Signed-off-by: Matthew Nickson <mnickson@sidingsmedia.com>