Switched to update
For some reason, if we try to run .save() after removing the last domain in a users list, the empty array is not saved and the last domain is not deleted. Domains can be deleted if there are more than one, however the very last domain could not.

A similar issue was seen when adding an alias, hence that has also been changed to .update()

The return types of the two methods have been changed, however they are only used in one place each and the return values are not used so there is no issue here.

Signed-off-by: Matthew Nickson <mnickson@sidingsmedia.com>